Tweak cost function for performance
The improvement is small on ditutte_fragment, maybe a few percent, but
larger on ditutte where the search reaches a score of 576 after 31
seconds instead of 45.

I did anticipate bigger gains but forgot to consider that each crossing
should be penalized only once requiring a bunch of set unions before
the intersection.